+++ include/rocksdb/options.h
@@ -1659,13 +1659,32 @@ struct WriteOptions {
   // Default: false
   bool memtable_insert_hint_per_batch;
 
+  // For automatic WAL flush (`Options::manual_wal_flush` == false)
+  // after the write operation associated with this option, charge
+  // the internal rate limiter (see `DBOptions::rate_limiter`)
+  // at this specified priority. The special value
+  // `Env::IO_TOTAL` disables charging the rate limiter.
+  //
+  // Only `Env::IO_USER` and `Env::IO_TOTAL` are allowed
+  // due to implementation constraints.
+  //
+  // If `Options::manual_wal_flush` == true or
+  // priority is set to ones other than
+  // `Env::IO_USER` and `Env::IO_TOTAL`,
+  // it has the same effect as setting
+  // rate_limiter_priority = `Env::IO_TOTAL`.
+  //
+  // Default: `Env::IO_TOTAL`
+  Env::IOPriority rate_limiter_priority;
+
   WriteOptions()
       : sync(false),
         disableWAL(false),
         ignore_missing_column_families(false),
         no_slowdown(false),
         low_pri(false),
-        memtable_insert_hint_per_batch(false) {}
+        memtable_insert_hint_per_batch(false),
+        rate_limiter_priority(Env::IO_TOTAL) {}
 };
 
 // Options that control flush operations